third_party/WebKit/LayoutTests/fast/ime/inputcontext-confirmcomposition-expected.txt

This tests confirmComposition() of InputMethodContext.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context instanceof InputMethodContext is true
Check if composition text 'abcd' is committed properly.
PASS compositionstart fired
PASS compositionend fired
PASS event.type is "compositionend"
PASS event.data is "abcd"
PASS textInput fired
PASS event.type is "textInput"
PASS event.data is "abcd"
PASS test.value is "abcd"
Check if no compositionend event will occur when composition is empty.
PASS test.value is "abcd"
Check if composition text 'efgh' is committed properly, and appended to the input field's value.
PASS compositionstart fired
PASS compositionend fired
PASS event.type is "compositionend"
PASS event.data is "efgh"
PASS textInput fired
PASS event.type is "textInput"
PASS event.data is "efgh"
PASS test.value is "abcdefgh"
Check if confirmComposition on non-focused element doesn't affect the focused editor.
PASS compositionstart fired
PASS compositionend fired
PASS event.type is "compositionend"
PASS event.data is ""
PASS test.value is "abcdefgh"
PASS successfullyParsed is true

TEST COMPLETE
 
